+++ src/components/Menu/Menu.stories.tsx
@@ -3,6 +3,7 @@ import React from "react";
 import { Story, Meta } from "@storybook/react/types-6-0";
 
 import { Menu as MenuComponent, MenuProps } from ".";
+import { MenuItem } from "../MenuItem";
 
 export default {
   title: "Components/Menu",
@@ -13,5 +14,10 @@ const Template: Story<MenuProps> = (args) => <MenuComponent {...args} />;
 
 export const Menu = Template.bind({});
 Menu.args = {
-  title: "Test"
+  title: "Test",
+  button: {
+    label: "Dismiss All",
+    onClick: alert,
+  },
+  content: <MenuItem />,
 };