+++ src/components/MenuItem/MenuItem.test.tsx
@@ -1,13 +1,12 @@
-import { render, screen } from '@testing-library/react'
-import '@testing-library/jest-dom/extend-expect'
+import { render, screen } from "@testing-library/react";
+import "@testing-library/jest-dom/extend-expect";
 
 import { MenuItem } from ".";
 
 test("Renders the component correctly", () => {
-    // Render component
-    render(<MenuItem />)
+  // Render component
+  render(<MenuItem />);
 
-    // Find the element
-    screen.getByTestId('menu-item')
-
+  // Find the element
+  screen.getByTestId("menu-item");
 });