+++ src/components/MenuItem/index.tsx
@@ -1,10 +1,9 @@
-export interface MenuItemProps {
-}
+export interface MenuItemProps {}
 
-export function MenuItem({ }: MenuItemProps) {
+export function MenuItem({}: MenuItemProps) {
   return (
     <div data-testid="menu-item">
       <p>Menu Item</p>
     </div>
-  )
+  );
 }