+++ src/containers/Reddit/Content/index.tsx
@@ -4,26 +4,32 @@ import { Menu, MenuItem, PostContent } from "../../../components";
 import { IRedditTop50DataItem } from "../../../constants/services/reddit/types";
 
 export interface ContentProps {
-    data: IRedditTop50DataItem[]
+  data: IRedditTop50DataItem[];
 }
 
 export default function Content(props: ContentProps) {
+  const { data } = props;
 
-    const { data } = props;
+  console.log(">>>>>data", data);
 
-    console.log(">>>>>data", data);
+  const onDismissAllHandler = () => {};
 
-    return (
-        <StyledContent>
-            <Menu title="Reddit Posts">
-                <MenuItem />
-            </Menu>
-            <PostContent />
-        </StyledContent>
-    )
+  return (
+    <StyledContent>
+      <Menu
+        title="Reddit Posts"
+        button={{
+          label: "Dismiss All",
+          onClick: onDismissAllHandler,
+        }}
+        content={<MenuItem />}
+      />
+      <PostContent />
+    </StyledContent>
+  );
 }
 
 const StyledContent = styled.div`
-    display: flex;
-    flex-direction: row;
+  display: flex;
+  flex-direction: row;
 `;